Split steps in to different modules.
